requirements.txt
pandas
rank-bm25
sentence-transformers
faiss-cpu
indic-transliteration
numpy